requirements.txt
anthropic==0.96.0
azure-storage-blob==12.28.0
groq==0.37.1
langchain-groq==0.3.2
mcp==1.27.0
numpy==2.2.6
pandas==2.3.3
plotly==6.7.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
streamlit==1.51.0
yfinance==1.3.0
